// convex/_generated/api.d.ts
declare const fullApi: ApiFromModules<{
  agregates: typeof agregates;
  auth: typeof auth;
  categories: typeof categories;
  defaultCategories: typeof defaultCategories;
  http: typeof http;
  movements: typeof movements;
  runs: typeof runs;
  settings: typeof settings;
}>;
declare const fullApiWithMounts: typeof fullApi;

export declare const api: FilterApi<
  typeof fullApiWithMounts,
  FunctionReference<any, "public">
>;
export declare const internal: FilterApi<
  typeof fullApiWithMounts,
  FunctionReference<any, "internal">
>;